refactor: Move `SeedFinderConfig` to `SeedFinder.hpp`
