+++ Core/include/Acts/Seeding/SeedFinderOrthogonal.hpp
@@ -10,22 +10,18 @@
 
 #include "Acts/EventData/SpacePointMutableData.hpp"
 #include "Acts/Seeding/SeedFilter.hpp"
-#include "Acts/Seeding/SeedFinderConfig.hpp"
+#include "Acts/Seeding/SeedFinder.hpp"
 #include "Acts/Seeding/SeedFinderOrthogonalConfig.hpp"
 #include "Acts/Utilities/KDTree.hpp"
 #include "Acts/Utilities/Logger.hpp"
 
 #include <array>
-#include <iostream>
-#include <list>
-#include <map>
 #include <memory>
-#include <set>
 #include <string>
-#include <utility>
 #include <vector>
 
 namespace Acts {
+
 template <typename external_spacepoint_t>
 class SeedFinderOrthogonal {
  public:
@@ -57,9 +53,9 @@ class SeedFinderOrthogonal {
    * @param logger The ACTS logger.
    */
   explicit SeedFinderOrthogonal(
-      const Acts::SeedFinderOrthogonalConfig<external_spacepoint_t> &config,
-      std::unique_ptr<const Acts::Logger> logger =
-          Acts::getDefaultLogger("Finder", Logging::Level::INFO));
+      const SeedFinderOrthogonalConfig<external_spacepoint_t> &config,
+      std::unique_ptr<const Logger> logger =
+          getDefaultLogger("Finder", Logging::Level::INFO));
   /**
    * @brief Destroy the orthogonal seed finder object.
    */
@@ -104,7 +100,7 @@ class SeedFinderOrthogonal {
    * covariance of the external space point
    */
   template <typename input_container_t, typename output_container_t>
-  void createSeeds(const Acts::SeedFinderOptions &options,
+  void createSeeds(const SeedFinderOptions &options,
                    const input_container_t &spacePoints,
                    output_container_t &out_cont) const;
 
@@ -124,7 +120,7 @@ class SeedFinderOrthogonal {
    * @return A vector of seeds.
    */
   template <typename input_container_t>
-  std::vector<seed_t> createSeeds(const Acts::SeedFinderOptions &options,
+  std::vector<seed_t> createSeeds(const SeedFinderOptions &options,
                                   const input_container_t &spacePoints) const;
 
  private:
@@ -214,8 +210,7 @@ class SeedFinderOrthogonal {
    * seed candidates to.
    */
   void filterCandidates(
-      const SeedFinderOptions &options,
-      Acts::SpacePointMutableData &mutableData,
+      const SeedFinderOptions &options, SpacePointMutableData &mutableData,
       const external_spacepoint_t &middle,
       const std::vector<const external_spacepoint_t *> &bottom,
       const std::vector<const external_spacepoint_t *> &top,
@@ -237,14 +232,14 @@ class SeedFinderOrthogonal {
    */
   template <typename output_container_t>
   void processFromMiddleSP(const SeedFinderOptions &options,
-                           Acts::SpacePointMutableData &mutableData,
+                           SpacePointMutableData &mutableData,
                            const tree_t &tree, output_container_t &out_cont,
                            const typename tree_t::pair_t &middle_p) const;
 
   /**
    * @brief The configuration for the seeding algorithm.
    */
-  Acts::SeedFinderOrthogonalConfig<external_spacepoint_t> m_config;
+  SeedFinderOrthogonalConfig<external_spacepoint_t> m_config;
 
   /**
    * @brief Get the logger.
@@ -254,8 +249,9 @@ class SeedFinderOrthogonal {
   /**
    * @brief The logger
    */
-  std::unique_ptr<const Acts::Logger> m_logger{getDummyLogger().clone()};
+  std::unique_ptr<const Logger> m_logger{getDummyLogger().clone()};
 };
+
 }  // namespace Acts
 
 #include "Acts/Seeding/SeedFinderOrthogonal.ipp"
