+++ Core/include/Acts/Seeding/detail/CylindricalSpacePointGrid.hpp
@@ -8,8 +8,9 @@
 
 #pragma once
 
+#include "Acts/Definitions/Units.hpp"
 #include "Acts/Seeding/BinnedGroup.hpp"
-#include "Acts/Seeding/SeedFinderConfig.hpp"
+#include "Acts/Seeding/SeedFinder.hpp"
 #include "Acts/Utilities/Grid.hpp"
 #include "Acts/Utilities/Logger.hpp"
 
@@ -29,43 +30,43 @@ concept CylindricalGridElement = requires(external_spacepoint_t sp) {
 
 /// Cylindrical Space Point bin is a 2D grid with (phi, z) bins
 /// It stores a vector of internal space points to external space points
-template <Acts::CylindricalGridElement external_spacepoint_t>
-using CylindricalSpacePointGrid = Acts::Grid<
-    std::vector<const external_spacepoint_t*>,
-    Acts::Axis<Acts::AxisType::Equidistant, Acts::AxisBoundaryType::Closed>,
-    Acts::Axis<Acts::AxisType::Variable, Acts::AxisBoundaryType::Open>,
-    Acts::Axis<Acts::AxisType::Variable, Acts::AxisBoundaryType::Open>>;
+template <CylindricalGridElement external_spacepoint_t>
+using CylindricalSpacePointGrid =
+    Grid<std::vector<const external_spacepoint_t*>,
+         Axis<AxisType::Equidistant, AxisBoundaryType::Closed>,
+         Axis<AxisType::Variable, AxisBoundaryType::Open>,
+         Axis<AxisType::Variable, AxisBoundaryType::Open>>;
 
 /// Cylindrical Binned Group
 template <typename external_spacepoint_t>
 using CylindricalBinnedGroup =
-    Acts::BinnedGroup<Acts::CylindricalSpacePointGrid<external_spacepoint_t>>;
+    BinnedGroup<CylindricalSpacePointGrid<external_spacepoint_t>>;
 
 template <typename external_spacepoint_t>
-using CylindricalBinnedGroupIterator = Acts::BinnedGroupIterator<
-    Acts::CylindricalSpacePointGrid<external_spacepoint_t>>;
+using CylindricalBinnedGroupIterator =
+    BinnedGroupIterator<CylindricalSpacePointGrid<external_spacepoint_t>>;
 
 struct CylindricalSpacePointGridConfig {
   // minimum pT to be found by seedFinder
-  float minPt = 0 * Acts::UnitConstants::MeV;
+  float minPt = 0 * UnitConstants::MeV;
   // maximum extension of sensitive detector layer relevant for seeding as
   // distance from x=y=0 (i.e. in r)
-  float rMax = 320 * Acts::UnitConstants::mm;
+  float rMax = 320 * UnitConstants::mm;
   // maximum extension of sensitive detector layer relevant for seeding as
   // distance from x=y=0 (i.e. in r)
-  float rMin = 0 * Acts::UnitConstants::mm;
+  float rMin = 0 * UnitConstants::mm;
   // maximum extension of sensitive detector layer relevant for seeding in
   // positive direction in z
-  float zMax = 0 * Acts::UnitConstants::mm;
+  float zMax = 0 * UnitConstants::mm;
   // maximum extension of sensitive detector layer relevant for seeding in
   // negative direction in z
-  float zMin = 0 * Acts::UnitConstants::mm;
+  float zMin = 0 * UnitConstants::mm;
   // maximum distance in r from middle space point to bottom or top spacepoint
-  float deltaRMax = 0 * Acts::UnitConstants::mm;
+  float deltaRMax = 0 * UnitConstants::mm;
   // maximum forward direction expressed as cot(theta)
   float cotThetaMax = 0;
   // maximum impact parameter in mm
-  float impactMax = 0 * Acts::UnitConstants::mm;
+  float impactMax = 0 * UnitConstants::mm;
   // minimum phi value for phiAxis construction
   float phiMin = -std::numbers::pi_v<float>;
   // maximum phi value for phiAxis construction
@@ -90,7 +91,7 @@ struct CylindricalSpacePointGridConfig {
           "CylindricalSpacePointGridConfig");
     }
 
-    using namespace Acts::UnitLiterals;
+    using namespace UnitLiterals;
     CylindricalSpacePointGridConfig config = *this;
     config.isInInternalUnits = true;
     config.minPt /= 1_MeV;
@@ -136,7 +137,7 @@ struct CylindricalSpacePointGridConfig {
 
 struct CylindricalSpacePointGridOptions {
   // magnetic field
-  float bFieldInZ = 0. * Acts::UnitConstants::T;
+  float bFieldInZ = 0. * UnitConstants::T;
   bool isInInternalUnits = false;
   CylindricalSpacePointGridOptions toInternalUnits() const {
     if (isInInternalUnits) {
@@ -144,7 +145,7 @@ struct CylindricalSpacePointGridOptions {
           "Repeated conversion to internal units for "
           "CylindricalSpacePointGridOptions");
     }
-    using namespace Acts::UnitLiterals;
+    using namespace UnitLiterals;
     CylindricalSpacePointGridOptions options = *this;
     options.isInInternalUnits = true;
     options.bFieldInZ /= 1000_T;
@@ -157,31 +158,29 @@ struct CylindricalSpacePointGridOptions {
 class CylindricalSpacePointGridCreator {
  public:
   template <typename external_spacepoint_t>
-  static Acts::CylindricalSpacePointGrid<external_spacepoint_t> createGrid(
-      const Acts::CylindricalSpacePointGridConfig& _config,
-      const Acts::CylindricalSpacePointGridOptions& _options,
-      const Acts::Logger& logger = Acts::getDummyLogger());
+  static CylindricalSpacePointGrid<external_spacepoint_t> createGrid(
+      const CylindricalSpacePointGridConfig& _config,
+      const CylindricalSpacePointGridOptions& _options,
+      const Logger& logger = getDummyLogger());
 
   template <typename external_spacepoint_t,
             typename external_spacepoint_iterator_t>
-  static void fillGrid(
-      const Acts::SeedFinderConfig<external_spacepoint_t>& config,
-      const Acts::SeedFinderOptions& options,
-      Acts::CylindricalSpacePointGrid<external_spacepoint_t>& grid,
-      external_spacepoint_iterator_t spBegin,
-      external_spacepoint_iterator_t spEnd,
-      const Acts::Logger& logger = Acts::getDummyLogger());
+  static void fillGrid(const SeedFinderConfig<external_spacepoint_t>& config,
+                       const SeedFinderOptions& options,
+                       CylindricalSpacePointGrid<external_spacepoint_t>& grid,
+                       external_spacepoint_iterator_t spBegin,
+                       external_spacepoint_iterator_t spEnd,
+                       const Logger& logger = getDummyLogger());
 
   template <typename external_spacepoint_t, typename external_collection_t>
     requires std::ranges::range<external_collection_t> &&
              std::same_as<typename external_collection_t::value_type,
                           external_spacepoint_t>
-  static void fillGrid(
-      const Acts::SeedFinderConfig<external_spacepoint_t>& config,
-      const Acts::SeedFinderOptions& options,
-      Acts::CylindricalSpacePointGrid<external_spacepoint_t>& grid,
-      const external_collection_t& collection,
-      const Acts::Logger& logger = Acts::getDummyLogger());
+  static void fillGrid(const SeedFinderConfig<external_spacepoint_t>& config,
+                       const SeedFinderOptions& options,
+                       CylindricalSpacePointGrid<external_spacepoint_t>& grid,
+                       const external_collection_t& collection,
+                       const Logger& logger = getDummyLogger());
 };
 
 }  // namespace Acts
