+++ Core/include/Acts/Seeding/detail/CylindricalSpacePointGrid.ipp
@@ -6,15 +6,13 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include <concepts>
-#include <numbers>
+namespace Acts {
 
 template <typename external_spacepoint_t>
-Acts::CylindricalSpacePointGrid<external_spacepoint_t>
-Acts::CylindricalSpacePointGridCreator::createGrid(
-    const Acts::CylindricalSpacePointGridConfig& config,
-    const Acts::CylindricalSpacePointGridOptions& options,
-    const Acts::Logger& logger) {
+CylindricalSpacePointGrid<external_spacepoint_t>
+CylindricalSpacePointGridCreator::createGrid(
+    const CylindricalSpacePointGridConfig& config,
+    const CylindricalSpacePointGridOptions& options, const Logger& logger) {
   if (!config.isInInternalUnits) {
     throw std::runtime_error(
         "CylindricalSpacePointGridConfig not in ACTS internal units in "
@@ -25,8 +23,8 @@ Acts::CylindricalSpacePointGridCreator::createGrid(
         "CylindricalSpacePointGridOptions not in ACTS internal units in "
         "CylindricalSpacePointGridCreator::createGrid");
   }
-  using AxisScalar = Acts::Vector3::Scalar;
-  using namespace Acts::UnitLiterals;
+  using AxisScalar = Vector3::Scalar;
+  using namespace UnitLiterals;
 
   int phiBins = 0;
   // for no magnetic field, create 100 phi-bins
@@ -102,7 +100,7 @@ Acts::CylindricalSpacePointGridCreator::createGrid(
     phiBins = std::min(phiBins, config.maxPhiBins);
   }
 
-  Acts::Axis<AxisType::Equidistant, AxisBoundaryType::Closed> phiAxis(
+  Axis<AxisType::Equidistant, AxisBoundaryType::Closed> phiAxis(
       config.phiMin, config.phiMax, phiBins);
 
   // vector that will store the edges of the bins of z
@@ -150,18 +148,18 @@ Acts::CylindricalSpacePointGridCreator::createGrid(
   ACTS_VERBOSE("- Z axis  : " << zAxis);
   ACTS_VERBOSE("- R axis  : " << rAxis);
 
-  return Acts::CylindricalSpacePointGrid<external_spacepoint_t>(
+  return CylindricalSpacePointGrid<external_spacepoint_t>(
       std::make_tuple(std::move(phiAxis), std::move(zAxis), std::move(rAxis)));
 }
 
 template <typename external_spacepoint_t,
           typename external_spacepoint_iterator_t>
-void Acts::CylindricalSpacePointGridCreator::fillGrid(
-    const Acts::SeedFinderConfig<external_spacepoint_t>& config,
-    const Acts::SeedFinderOptions& options,
-    Acts::CylindricalSpacePointGrid<external_spacepoint_t>& grid,
+void CylindricalSpacePointGridCreator::fillGrid(
+    const SeedFinderConfig<external_spacepoint_t>& config,
+    const SeedFinderOptions& options,
+    CylindricalSpacePointGrid<external_spacepoint_t>& grid,
     external_spacepoint_iterator_t spBegin,
-    external_spacepoint_iterator_t spEnd, const Acts::Logger& logger) {
+    external_spacepoint_iterator_t spEnd, const Logger& logger) {
   if (!config.isInInternalUnits) {
     throw std::runtime_error(
         "SeedFinderConfig not in ACTS internal units in BinnedSPGroup");
@@ -201,7 +199,7 @@ void Acts::CylindricalSpacePointGridCreator::fillGrid(
     }
 
     // fill rbins into grid
-    Acts::Vector3 position(sp.phi(), sp.z(), sp.radius());
+    Vector3 position(sp.phi(), sp.z(), sp.radius());
     if (!grid.isInside(position)) {
       continue;
     }
@@ -233,12 +231,14 @@ template <typename external_spacepoint_t, typename external_collection_t>
   requires std::ranges::range<external_collection_t> &&
            std::same_as<typename external_collection_t::value_type,
                         external_spacepoint_t>
-void Acts::CylindricalSpacePointGridCreator::fillGrid(
-    const Acts::SeedFinderConfig<external_spacepoint_t>& config,
-    const Acts::SeedFinderOptions& options,
-    Acts::CylindricalSpacePointGrid<external_spacepoint_t>& grid,
-    const external_collection_t& collection, const Acts::Logger& logger) {
-  Acts::CylindricalSpacePointGridCreator::fillGrid<external_spacepoint_t>(
+void CylindricalSpacePointGridCreator::fillGrid(
+    const SeedFinderConfig<external_spacepoint_t>& config,
+    const SeedFinderOptions& options,
+    CylindricalSpacePointGrid<external_spacepoint_t>& grid,
+    const external_collection_t& collection, const Logger& logger) {
+  CylindricalSpacePointGridCreator::fillGrid<external_spacepoint_t>(
       config, options, grid, std::ranges::begin(collection),
       std::ranges::end(collection), logger);
 }
+
+}  // namespace Acts
