+++ Plugins/Cuda/include/Acts/Plugins/Cuda/Seeding/SeedFinder.hpp
@@ -14,7 +14,6 @@
 #include "Acts/Seeding/InternalSpacePoint.hpp"
 #include "Acts/Seeding/SeedFilter.hpp"
 #include "Acts/Seeding/SeedFinder.hpp"
-#include "Acts/Seeding/SeedFinderConfig.hpp"
 #include "Acts/Seeding/SpacePointGrid.hpp"
 
 #include <array>
