+++ Plugins/Cuda/include/Acts/Plugins/Cuda/Seeding2/SeedFinder.hpp
@@ -15,7 +15,6 @@
 #include "Acts/EventData/SpacePointData.hpp"
 #include "Acts/Seeding/Seed.hpp"
 #include "Acts/Seeding/SeedFilterConfig.hpp"
-#include "Acts/Seeding/SeedFinderConfig.hpp"
 #include "Acts/Seeding/SpacePointGrid.hpp"
 #include "Acts/Utilities/Logger.hpp"
